Add DELETE and POST requests for offers
