+++ boilerplate/boilerplate/detailed_views.py
@@ -1,12 +1,9 @@
+from django.http import Http404
 from rest_framework import generics as gr
+from rest_framework.response import Response
 
 from boilerplate.models import *
 from boilerplate.serializers import *
-
-
-class ClientsListView(gr.ListAPIView):
-    queryset = Client.objects.all()
-    serializer_class = ClientSerializer
 
 
 class ClientDetailView(gr.RetrieveAPIView):
@@ -15,30 +12,9 @@ class ClientDetailView(gr.RetrieveAPIView):
     lookup_field = 'id'
 
 
-class SourcesListView(gr.ListAPIView):
-    queryset = Source.objects.all()
-    serializer_class = SourceSerializer
-
-
 class SourceDetailView(gr.RetrieveAPIView):
     queryset = Source.objects.all()
     serializer_class = SourceSerializer
-    lookup_field = 'id'
-
-
-class OffersListView(gr.ListAPIView):
-    queryset = Offer.objects.all()
-    serializer_class = OfferSerializer
-
-
-class DealersListView(gr.ListAPIView):
-    queryset = Dealer.objects.all()
-    serializer_class = DealerSerializer
-
-
-class OfferDetailView(gr.RetrieveAPIView):
-    queryset = Offer.objects.all()
-    serializer_class = OfferSerializer
     lookup_field = 'id'
 
 
@@ -54,9 +30,21 @@ class ClientDetailView(gr.RetrieveAPIView):
     lookup_field = 'id'
 
 
-class OffersOrdersListView(gr.ListCreateAPIView):
-    queryset = OfferOrder.objects.all()
-    serializer_class = OfferOrderSerializer
+class OfferDetailView(gr.RetrieveDestroyAPIView):
+    queryset = Offer.objects.all()
+    serializer_class = OfferSerializer
+    lookup_field = 'id'
+
+    def destroy(self, request, *args, **kwargs):
+        try:
+            instance = self.get_object()
+        except Http404:
+            return Response({'result': False})
+        serializer = self.get_serializer(instance)
+        data = serializer.data
+        data['result'] = True
+        self.perform_destroy(instance)
+        return Response(data)
 
 
 class OfferOrderDetailView(gr.RetrieveAPIView):
